Refactored structure of order form
