+++ src/components/App.scss
@@ -38,11 +38,11 @@
   // max-width: 1440px;
 }
 
-.order-content {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  @media screen and (max-width: $tablet) {
-    flex-direction: column;
-  }
-}
+// .order-content {
+//   display: flex;
+//   width: 100%;
+//   height: 100%;
+//   @media screen and (max-width: $tablet) {
+//     flex-direction: column;
+//   }
+// }
