+++ src/components/MainContent/MainContent.js
@@ -9,7 +9,7 @@ function MainContent() {
       <h1 className="main-content__title main-content__title_rus">Каршеринг</h1>
       <h2 className="main-content__title main-content__title_en" lang="en">Need for drive</h2>
       <p className="main-content__text">Поминутная аренда авто твоего города</p>
-      <Link to="/location">
+      <Link to="/order-form/location">
         <Button 
           style={{ width: 250, marginTop: 64 }}
           text="Забронировать"
