+++ src/components/OrderForm/OrderForm.scss
@@ -1,0 +1,10 @@
+@import "../../scss/variables.scss";
+
+.order-form {
+  display: flex;
+  width: 100%;
+  height: 100%;
+  @media screen and (max-width: $tablet) {
+    flex-direction: column;
+  }
+}
