+++ src/components/OrderForm/components/Order/Order.scss
@@ -1,4 +1,4 @@
-@import "../../scss/variables.scss";
+@import "../../../../scss/variables.scss";
 
 .order {
   min-width: 384px;
